requirements.txt
pytest
flask
gitpython
flask-restful
Flask-SQLAlchemy
python-dotenv
requests